Beginning work on more accurate serial emulation + external serial interface.
Currently only emulates the single speed dmg serial port, cgb speeds not
yet implemented. More importantly, since serial communication is now done
bit by bit, there needs to be a way to interrupt the emulator mid frame
when an external device is driving the serial clock and wants to shift a
bit in.