The bits in joyp for the A and B key were in the wrong order.
Key::A is supposed to be bit 0, and Key::B is supposed to be bit 1 when
reading from JOYP with the button keys selected.
Also gave the Key enum explicit values so they can be used as indices,
as well as marking them repr(C) so they can be used directly in a C api
instead of being converted to integers and back.